requirements.txt
beautifulsoup4==4.14.2
fastapi==0.118.0
fastapi-jwt==0.3.0
sqlite4==0.1.1
pandas==2.3.3
requests==2.32.5
uvicorn==0.37.0
sqlalchemy==2.0.43
python-jose[cryptography]==3.5.0
passlib==1.7.4 # Forcing bcrypt to 4.3.0 due to issue below
bcrypt==4.3.0 # https://foss.heptapod.net/python-libs/passlib/-/issues/196
pydantic[email]==2.11.10
python-dotenv==1.1.1
psycopg[binary,pool]==3.2.10
